Prueba 3

+++ main.go
@@ -3,8 +3,8 @@ package main
 import "fmt"
 
 func main() {
-	var lado float32
-	fmt.Scanf("%f", &lado)
+	var lado int
+	fmt.Scanf("%d", &lado)
 	area := lado * lado
 	fmt.Println(area)
 }
